refactor: restructure modules into folders (#12)
